Change the total and price to decimal to avoid rounding issues

+++ src/OrderApp.Api/DataTransferObjects/OrderDetailsDto.cs
@@ -7,4 +7,4 @@ namespace OrderApp.Api.DataTransferObjects;
 /// <param name="Items">The items in the order.</param>
 /// <param name="Total">The total cost.</param>
 /// <param name="Classification">The classification of the order.</param>
-public sealed record OrderDetailsDto(string CustomerId, OrderItemsDto[] Items, double Total, string Classification);
+public sealed record OrderDetailsDto(string CustomerId, OrderItemsDto[] Items, decimal Total, string Classification);
