+++ src/OrderApp.Query/Helpers/IOrderClassificationHelper.cs
@@ -10,5 +10,5 @@ public interface IOrderClassificationHelper
     /// </summary>
     /// <param name="orderTotal">The order total.</param>
     /// <returns>The order classification.</returns>
-    string GetOrderClassification(double orderTotal);
+    string GetOrderClassification(decimal orderTotal);
 }
